[MOGURI-8] add: GoalQuest 목표 퀘스트 리스트 조회 추가

+++ moguri/src/main/java/org/moguri/security/account/domain/MemberVO.java
@@ -14,7 +14,7 @@ import java.math.BigDecimal;
 @AllArgsConstructor
 @Builder
 public class MemberVO {
-    private int memberId; // MEMBER_ID를 PK로 사용
+    private long memberId; // MEMBER_ID를 PK로 사용
     private String email; // 이메일
     private String nickName; // 닉네임
     private String password; // 비밀번호
